move todo type difinition to models
