+++ models/todo.ts
@@ -1,4 +1,14 @@
-import { ITodoData, ITodo } from '~/types/todo'
+export interface ITodoData {
+  uid: string
+  content: string
+  priority: number
+  done: boolean
+  doneAt: firebase.firestore.FieldValue | null
+}
+
+export interface ITodo extends ITodoData {
+  id: string
+}
 
 export class Todo implements ITodo {
   id: string
