--- types/todo.ts
@@ -1,11 +1,0 @@
-export interface ITodoData {
-  uid: string
-  content: string
-  priority: number
-  done: boolean
-  doneAt: firebase.firestore.FieldValue | null
-}
-
-export interface ITodo extends ITodoData {
-  id: string
-}
